refactor(github): use dynamic import for subcommand
