+++ packages/opencode/src/cli/cmd/github/util.ts
@@ -1,0 +1,41 @@
+import type { MessageV2 } from "../../../session/message-v2"
+
+// Parses GitHub remote URLs in various formats:
+// - https://github.com/owner/repo.git
+// - https://github.com/owner/repo
+// - git@github.com:owner/repo.git
+// - git@github.com:owner/repo
+// - ssh://git@github.com/owner/repo.git
+// - ssh://git@github.com/owner/repo
+export function parseGitHubRemote(url: string): { owner: string; repo: string } | null {
+  const match = url.match(/^(?:(?:https?|ssh):\/\/)?(?:git@)?github\.com[:/]([^/]+)\/([^/]+?)(?:\.git)?$/)
+  if (!match) return null
+  return { owner: match[1], repo: match[2] }
+}
+
+/**
+ * Extracts displayable text from assistant response parts.
+ * Returns null for non-text responses (signals summary needed).
+ * Throws only for truly empty responses.
+ */
+export function extractResponseText(parts: MessageV2.Part[]): string | null {
+  const textPart = parts.findLast((p) => p.type === "text")
+  if (textPart) return textPart.text
+
+  // Non-text parts (tools, reasoning, step-start/step-finish, etc.) - signal summary needed
+  if (parts.length > 0) return null
+
+  throw new Error("Failed to parse response: no parts returned")
+}
+
+/**
+ * Formats a PROMPT_TOO_LARGE error message with details about files in the prompt.
+ * Content is base64 encoded, so we calculate original size by multiplying by 0.75.
+ */
+export function formatPromptTooLargeError(files: { filename: string; content: string }[]): string {
+  const fileDetails =
+    files.length > 0
+      ? `\n\nFiles in prompt:\n${files.map((f) => `  - ${f.filename} (${((f.content.length * 0.75) / 1024).toFixed(0)} KB)`).join("\n")}`
+      : ""
+  return `PROMPT_TOO_LARGE: The prompt exceeds the model's context limit.${fileDetails}`
+}
