+++ packages/opencode/src/index.ts
@@ -19,7 +19,7 @@ import { Filesystem } from "@/util/filesystem"
 import { DebugCommand } from "./cli/cmd/debug/command"
 import { StatsCommand } from "./cli/cmd/stats"
 import { McpCommand } from "./cli/cmd/mcp"
-import { GithubCommand } from "./cli/cmd/github"
+import { GithubCommand } from "./cli/cmd/github/command"
 import { ExportCommand } from "./cli/cmd/export/command"
 import { ImportCommand } from "./cli/cmd/import"
 import { AttachCommand } from "./cli/cmd/tui/attach"
