+++ packages/opencode/test/cli/github-remote.test.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "bun:test"
-import { parseGitHubRemote } from "../../src/cli/cmd/github"
+import { parseGitHubRemote } from "../../src/cli/cmd/github/util"
 
 test("parses https URL with .git suffix", () => {
   expect(parseGitHubRemote("https://github.com/sst/opencode.git")).toEqual({ owner: "sst", repo: "opencode" })
